chore: Avoid casting
(cherry picked from commit 3dbd6aeb6d07f8e66e81ec359cb3373fc91b05a0)

+++ src/Uno.UI.Runtime.Skia.WebAssembly.Browser/Devices/Input/BrowserPointerInputSource.cs
@@ -150,7 +150,7 @@ internal unsafe partial class BrowserPointerInputSource : IUnoCorePointerInputSo
 					throw new ArgumentOutOfRangeException(nameof(@event), $"Unknown event ({@event}-{evt}).");
 			}
 
-			return (byte)(args.Handled ? HtmlEventDispatchResult.PreventDefault : HtmlEventDispatchResult.Ok);
+			return args.Handled ? HtmlEventDispatchResult.PreventDefault : HtmlEventDispatchResult.Ok;
 		}
 		catch (Exception error)
 		{
